Move shared feeds to new rail and side navigation
- Create new rail item for shared feed home
- Remove shared feeds from the Tipline navigation list
- Add a new icon to the sidebar that opens a new side bar
- Add counter to the section title (number of shared feeds)
- Extract FeedsComponent from ProjectsComponent

Reference: CV2-3927
