+++ localization/react-intl/src/app/components/drawer/Projects/FeedsComponent.json
@@ -1,0 +1,22 @@
+[
+  {
+    "id": "projectsComponent.pendingInvitationFeedTitle",
+    "description": "Tooltip for a navigation item that has a status of a pending invitation",
+    "defaultMessage": "Pending invitation: {feedTitle}"
+  },
+  {
+    "id": "projectsComponent.sharedFeeds",
+    "description": "Feeds of content shared across workspaces",
+    "defaultMessage": "Colaborating [{feedsLength}]"
+  },
+  {
+    "id": "projectsComponent.newSharedFeed",
+    "description": "Tooltip for the button that navigates to shared feed creation page",
+    "defaultMessage": "New shared feed"
+  },
+  {
+    "id": "projectsComponent.noSharedFeeds",
+    "description": "Displayed under the shared feed header when there are no feeds in it",
+    "defaultMessage": "No shared feeds"
+  }
+]
