+++ src/app/components/drawer/Projects/ProjectsComponent.js
@@ -1,13 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage, injectIntl, defineMessages } from 'react-intl';
-import { browserHistory, withRouter, Link } from 'react-router';
+import { FormattedMessage, injectIntl } from 'react-intl';
+import { withRouter, Link } from 'react-router';
 import Collapse from '@material-ui/core/Collapse';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import ListItemText from '@material-ui/core/ListItemText';
 import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
-import cx from 'classnames/bind';
 import ProjectsListItem from './ProjectsListItem';
 import NewProject from './NewProject';
 import ButtonMain from '../../cds/buttons-checkboxes-chips/ButtonMain';
@@ -26,7 +25,6 @@ import UnmatchedIcon from '../../../icons/unmatched.svg';
 import Can from '../../Can';
 import DeleteIcon from '../../../icons/delete.svg';
 import ReportIcon from '../../../icons/report.svg';
-import ScheduleSendIcon from '../../../icons/schedule_send.svg';
 import { withSetFlashMessage } from '../../FlashMessage';
 import { assignedToMeDefaultQuery } from '../../team/AssignedToMe';
 import { suggestedMatchesDefaultQuery } from '../../team/SuggestedMatches';
@@ -37,21 +35,11 @@ import { tiplineInboxDefaultQuery } from '../../team/TiplineInbox';
 import ProjectsCoreListCounter from './ProjectsCoreListCounter';
 import styles from './Projects.module.css';
 
-const messages = defineMessages({
-  pendingInvitationFeedTooltip: {
-    id: 'projectsComponent.pendingInvitationFeedTitle',
-    defaultMessage: 'Pending invitation: {feedTitle}',
-    description: 'Tooltip for a navigation item that has a status of a pending invitation',
-  },
-});
-
 const ProjectsComponent = ({
   currentUser,
   team,
   savedSearches,
-  feeds,
   location,
-  intl,
 }) => {
   const [showNewListDialog, setShowNewListDialog] = React.useState(false);
   const [collapsed, setCollapsed] = React.useState(false);
@@ -63,8 +51,6 @@ const ProjectsComponent = ({
 
   const [listsExpanded, setListsExpanded] =
     React.useState(getBooleanPref('drawer.listsExpanded', true));
-  const [feedsExpanded, setFeedsExpanded] =
-    React.useState(getBooleanPref('drawer.feedsExpanded', true));
 
   // Get/set which list item should be highlighted
   const pathParts = window.location.pathname.split('/');
@@ -79,10 +65,6 @@ const ProjectsComponent = ({
 
   const handleAllItems = () => {
     setActiveItem({ type: 'all-items', id: null });
-  };
-
-  const handleCreateFeed = () => {
-    browserHistory.push(`/${team.slug}/feed/create`);
   };
 
   const handleSpecialLists = (listId) => {
@@ -101,11 +83,6 @@ const ProjectsComponent = ({
   const handleToggleListsExpand = () => {
     setListsExpanded(!listsExpanded);
     window.storage.set('drawer.listsExpanded', !listsExpanded);
-  };
-
-  const handleToggleFeedsExpand = () => {
-    setFeedsExpanded(!feedsExpanded);
-    window.storage.set('drawer.feedsExpanded', !feedsExpanded);
   };
 
   const handleTrash = () => {
@@ -335,83 +312,6 @@ const ProjectsComponent = ({
             }
           </Collapse>
         </React.Fragment>
-
-        {/* Shared feeds */}
-        <ListItem onClick={handleToggleFeedsExpand} className={[styles.listHeader, 'project-list__header'].join(' ')}>
-          { feedsExpanded ? <ExpandLessIcon className={styles.listChevron} /> : <ExpandMoreIcon className={styles.listChevron} /> }
-          <ListItemText disableTypography className={styles.listHeaderLabel}>
-            <FormattedMessage tagName="span" id="projectsComponent.sharedFeeds" defaultMessage="Shared feeds" description="Feeds of content shared across workspaces" />
-            <Can permissions={team.permissions} permission="create Feed">
-              <Tooltip arrow title={<FormattedMessage id="projectsComponent.newSharedFeed" defaultMessage="New shared feed" description="Tooltip for the button that navigates to shared feed creation page" />}>
-                <span className={styles.listHeaderLabelButton}>
-                  <ButtonMain
-                    className="projects-list__add-feed"
-                    iconCenter={<AddCircleIcon />}
-                    variant="contained"
-                    size="small"
-                    theme="text"
-                    onClick={(e) => { handleCreateFeed(); e.stopPropagation(); }}
-                  />
-                </span>
-              </Tooltip>
-            </Can>
-          </ListItemText>
-        </ListItem>
-        <Collapse in={feedsExpanded} className={styles.listCollapseWrapper}>
-          { feeds.length === 0 ?
-            <ListItem className={[styles.listItem, styles.listItem_containsCount, styles.listItem_empty].join(' ')}>
-              <ListItemText disableTypography className={styles.listLabel}>
-                <span>
-                  <FormattedMessage tagName="em" id="projectsComponent.noSharedFeeds" defaultMessage="No shared feeds" description="Displayed under the shared feed header when there are no feeds in it" />
-                </span>
-              </ListItemText>
-            </ListItem> :
-            <>
-              {feeds.sort((a, b) => (a?.title?.localeCompare(b.title))).map((feed) => {
-                let itemProps = {};
-                let itemType = null;
-                let itemIcon = null;
-                switch (feed.type) {
-                // Feeds created by the workspace
-                case 'Feed':
-                  itemProps = { routePrefix: 'feed' };
-                  itemType = 'feed';
-                  break;
-                // Feeds not created by the workspace, but joined upon invitation
-                case 'FeedTeam':
-                  itemProps = { routePrefix: 'feed' };
-                  itemType = 'feed-team';
-                  break;
-                // Feed invitations received but not processed yet
-                case 'FeedInvitation':
-                  itemProps = { routePrefix: 'feed', routeSuffix: '/invitation' };
-                  itemType = 'feed-invitation';
-                  itemIcon = <ScheduleSendIcon className={cx(styles.listIcon, styles.listIconInvitedFeed)} />;
-                  break;
-                default:
-                  break;
-                }
-                return (
-                  <ProjectsListItem
-                    className={cx(
-                      {
-                        [styles.listItemInvited]: feed.type === 'FeedInvitation',
-                      })
-                    }
-                    key={feed.id}
-                    project={feed}
-                    teamSlug={team.slug}
-                    onClick={handleClick}
-                    isActive={isActive(itemType, feed.dbid)}
-                    icon={itemIcon}
-                    tooltip={feed.type === 'FeedInvitation' ? intl.formatMessage(messages.pendingInvitationFeedTooltip, { feedTitle: feed.title }) : feed.title}
-                    {...itemProps}
-                  />
-                );
-              })}
-            </>
-          }
-        </Collapse>
       </List>
 
       <List dense disablePadding className={[styles.listWrapper, styles.listFooter].join(' ')}>
@@ -494,11 +394,6 @@ ProjectsComponent.propTypes = {
     dbid: PropTypes.number.isRequired,
     title: PropTypes.string.isRequired,
     filters: PropTypes.string.isRequired,
-  }).isRequired).isRequired,
-  feeds: PropTypes.arrayOf(PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    dbid: PropTypes.number.isRequired,
-    name: PropTypes.string.isRequired,
   }).isRequired).isRequired,
 };
